folder 'Input' renamed to 'Connector'
